created forms.py and a new db for the forms and also worked on the files upload and passcode
